minor edit, fixing documentation

+++ Breakneck Brigade/SousChef/SoundThing.cs
@@ -9,31 +9,33 @@ using System.Windows.Media;
 
 namespace SousChef
 {
-    /**
-     * Plays a sound.  Name was chosen by ET.  Usage:
-     * 
-     * 1) Add the sound file you'd like to include in the game to (supported types: wma, mp3, wav) "res\sounds\".
-     * 2) Let's say that your file is titled 'scratch.wav'.  Method call then, is SoundThing.Play("scratch").  The key *is* case sensitive.
-     * 3) Rebuild project & listen to music.
-     * 
-     * Method is case sensitive, and omits the file extension.  Files to be played MUST be in "res\sounds\"
-     * 
-     * @author Alec
-     */
+    /// <summary>
+    /// Plays a sound.  Name was chosen by ET.  Usage:
+    ///1) Add the sound file you'd like to include in the game to (supported types: wma, mp3, wav) "res\sounds\".
+    ///2) Let's say that your file is titled 'scratch.wav'.  Method call then, is SoundThing.Play("scratch").  The key *is* case sensitive.
+    ///3) Rebuild project & listen to beautiful music.
+    ///
+    /// Method is case sensitive, and omits the file extension.  Files to be played MUST be in "res\sounds\"
+    /// 
+    /// </summary>
     public class SoundThing
     {
-        /**
-         * Internal dictionary/hashmap which links files with their shorthand names.
-         */
+        /// <summary>
+        /// Internal dictionary/hashmap which links files with their shorthand names.
+        /// </summary>
         private static Dictionary<string, string> _l;
+
+        /// <summary>
+        /// I have no idea what this does, but Alex says we need it >.>
+        /// </summary>
         private static Dictionary<string, string> l { get { return _l ?? (_l = makeDict()); } }
 
-        /**
-         * Plays a sound. See class header comments for details.
-         * 
-         * @param key The key of the sound to play.  Note that this *is* case-sensitive.
-         */
-        public static void Play(BBSound key, int volume)
+        /// <summary>
+        /// Plays a sound. See class header comments for details.
+        /// </summary>
+        /// <param name="key">The key of the sound to play.  Note that this *is* case-sensitive.</param>
+        /// <param name="volume">The volume to play at, on a scale of 0-1, where 0 is silent and 1 is loudest.</param>
+        public static void Play(BBSound key, double volume)
         {
             String temp = "";
             if (l.TryGetValue(key.ToString(), out temp))
@@ -42,9 +44,10 @@ namespace SousChef
                 throw new Exception("no sound file for " + key.ToString());
         }
 
-        /**
-         * Generates our dictionary at runtime.
-         */
+        /// <summary>
+        /// Generates our dictionary at runtime.
+        /// </summary>
+        /// <returns>The dictionary linking each keys to file paths</returns>
         private static Dictionary<string, string> makeDict()
         {
             Dictionary<string, string> m = new Dictionary<string, string>();
@@ -60,38 +63,38 @@ namespace SousChef
             return m;
         }
 
-        /**
-         * Internal class representing a thread playing a sound.
-         * 
-         * @author Alec
-         */
+        /// <summary>
+        /// Internal class representing a thread playing a sound.
+        /// </summary>
         private class SoundThread
         {
-            /**
-             * The path to the file to play
-             */
+            /// <summary>
+            /// The path to the file to play
+            /// </summary>
             private string path;
 
-            /**
-             * The volume to play the sound at
-             */
+            /// <summary>
+            /// The volume to play the sound at
+            /// </summary>
             private double volume;
 
-            /**
-             * Constructor, creates a SoundThread with a specified sound to play.
-             * 
-             * @param path The path to the file to play
-             * @param volume The volume level to play a sound between 0-1, where 0 is queit, and 1 is loudest.
-             */
+
+            /// <summary>
+            /// Constructor, creates a SoundThread with a specified sound to play.
+            /// 
+            /// </summary>
+            /// <param name="path">The path to the file to play</param>
+            /// <param name="volume">The volume level to play a sound between 0-1, where 0 is queit, and 1 is loudest.</param>
             public SoundThread(string path, double volume)
             {
                 this.path = path;
                 this.volume = volume;
             }
+            
+             /// <summary>
+             /// Actually plays the sound.  Should only be called by Thread.
+             /// </summary>
 
-            /**
-             * Actually plays the sound.  Should only be called by Thread.
-             */
             public void DoWork()
             {
                 MediaPlayer player = new MediaPlayer();
